requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
APScheduler==3.6.3
arrow==1.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
black==23.12.0
blessed==1.20.0
cachetools==4.2.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.7
Django==4.2
django-celery-beat==2.5.0
django-enumfields==2.1.1
django-picklefield==3.1
django-q==1.3.9
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
flake8==6.1.0
flake8-annotations==3.0.1
flake8-quotes==3.3.2
flake8-variables-names==0.0.6
flower==2.0.1
frozenlist==1.4.0
humanize==4.9.0
idna==3.6
inflection==0.5.1
isort==5.13.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.3.4
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pep8-naming==0.13.3
Pillow==10.1.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-telegram-bot==13.15
pytz==2023.3.post1
PyYAML==6.0.1
redis==3.5.3
referencing==0.32.0
requests==2.31.0
rpds-py==0.15.2
six==1.16.0
sqlparse==0.4.4
stripe==7.8.2
tomli==2.0.1
tornado==6.1
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
yarl==1.9.4
